Do not compile with -Werror
-Werror is "turn warnings into errors". As compilers' diagnostic
capabilities improve, new warnings get added. This means that code can
fail to compile with newer compilers. We've seen this with clang 13,
and -Wdeprecated-copy.

Unfortunately, ignoring the warning is not as simple, as we some
compilers emit -Wunknown-warning-option/-Wunknown-warning for unkown
warning options, which themselves get turned into errors. We could also
work around that by not treating "unknown warning option" as an error,
or by detecting compiler versions.

We take the simple solution here, which is to not compile with -Werror
in the first place.